+++ src/Config.cpp
@@ -22,32 +22,6 @@ static const int DisableHtslibLogging = []() {
     return 0;
 }();
 
-bool DoesHtslibSupportLongCigar()
-{
-    const std::string htsVersion = hts_version();
-
-    // remove any "-<blah>" for non-release versions
-    const auto versionBase = PacBio::BAM::Split(htsVersion, '-');
-    if (versionBase.empty()) {
-        throw std::runtime_error{"[pbbam] config ERROR: invalid htslib version format: '" +
-                                 htsVersion + "'"};
-    }
-
-    // grab major/minor version numbers
-    const auto versionParts = PacBio::BAM::Split(versionBase[0], '.');
-    if (versionParts.size() < 2) {
-        throw std::runtime_error{"[pbbam] config ERROR: invalid htslib version format: '" +
-                                 htsVersion + "'"};
-    }
-
-    // check against v1.7
-    const int versionMajor = std::stoi(versionParts[0]);
-    const int versionMinor = std::stoi(versionParts[1]);
-    constexpr int V17_MAJOR = 1;
-    constexpr int V17_MINOR = 7;
-    return std::tie(versionMajor, versionMinor) >= std::tie(V17_MAJOR, V17_MINOR);
-}
-
 #ifdef PBBAM_PERMISSIVE_CIGAR
 static const bool PermissiveCigar = []() {
     Data::CigarOperation::DisableAutoValidation();